Bake in the Syreen Voice Fix text and fix Orz convo crash
The Syreen Voice Fix text will now be used when opting to disable the 3DO voices, specifically when talking about the Deep Child Egg Case Fragments, basically reverting it to the true PC version text.

And added in the necessary characters to the Orz comm text to activate the font switch for the *special* words. Omitting these characters crashed the game at random or when you went to review the conversation.